requirements.txt
requests
robotframework
robotframework-requests
robotframework-debuglibrary